Delete unnecessary compile options
As these compile options are machine specific, there is certanly no need to force their usage.

This also improves the CI pipeline with FlexRIC by minimizing the number of warnings.
In addition, option -march=native forces the use of e.g. avx512 which prevents the installation
of Service Models in the RAN image to be the same as for FlexRIC image, as they are built on
different machines.